simple mapping for shared address space

+++ libcommon/include/remote_pool.h
@@ -12,7 +12,7 @@ extern "C"
 
 /* sbuf starting from base_sbuf is occupied by remote pools */
 void manager_init();
-void add_pool(int pid, uint64_t linesize);
+void add_pool(int pid, unsigned linesize);
 
 void process_cache_req(RPC_rrf_t *req);
 void process_call_req(RPC_rrf_t *req);
